Move TarArchives in dedicated package
